feat: Add health_dietary_caffeine metric
For the moment this is an HTTP server with a `/metrics` endpoint that
Prometheus can hit. It correctly reports a metric which is the current
estimated amount of caffeine in my body based on data points of
individual caffeine consumption events. However, there is not yet an
ingestion endpoint to allow anything to easily record new records.